Renamed AuthModal components, updated Header and HeroSection
