+++ packages/ui/src/components/auth/SignForm.module.css
@@ -1,0 +1,120 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: 400px;
+  margin: 0 auto;
+}
+
+.title {
+  text-align: center;
+  font-size: 1.8rem;
+  color: var(--black);
+}
+
+.signintitle{
+margin-top: 8rem;
+}
+
+
+
+.signbutton {
+  margin: 1rem 0;
+}
+
+.submitButton:disabled {
+  background-color: #ccc;
+  cursor: not-allowed;
+}
+
+.errorMessage,
+.successMessage {
+  text-align: center;
+  width: 100%;
+  max-width: 400px;
+  font-size: 1rem;
+  color: #ff3333;
+}
+
+.successMessage {
+  color: green;
+}
+
+.inputActiontext {
+  font-size: 1rem;
+  color: var(--description);
+  font-weight: 500;
+  text-align: center;
+  margin-top: 10px;
+}
+
+.inputActiontext .signIn {
+  color: #007bff;
+  cursor: pointer;
+}
+
+.inputActiontext .signIn:hover {
+  text-decoration: underline;
+}
+
+.forgotPassword {
+  margin-left: 6rem;
+  color: #4b4bb5;
+}
+
+.cusomeinputcss {
+  width: 247px;
+}
+
+
+
+
+
+
+/* Mobile specific adjustments */
+@media (max-width: 768px) {
+  .form {
+    width: 90%; /* Adjust form width for mobile */
+    padding: 1rem; /* Padding for better mobile design */
+  }
+
+ 
+
+  
+
+  .submitButton {
+    width: 100%; /* Full-width button on mobile */
+    padding: 12px;
+  }
+
+ 
+
+  .forgotPassword {
+    margin-left: 0; /* Remove left margin for mobile */
+    text-align: center; /* Center the "Forgot Password?" link */
+  }
+
+  .inputActiontext {
+    font-size: 0.9rem; /* Smaller text for action links */
+  }
+}
+
+@media (max-width: 480px) {
+  .form {
+    width: 90%; /* Full width for form on mobile */
+    padding: 15px;
+  }
+
+  
+
+  .submitButton {
+    font-size: 16px;
+    padding: 10px;
+  }
+
+  
+
+  .inputActiontext {
+    font-size: 0.9rem; /* Slightly smaller font for action texts */
+  }
+}
