--- packages/ui/src/components/authmodal/SignForm.module.css
@@ -1,84 +1,0 @@
-
-
-
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  max-width: 400px;
-  margin: 0 auto;
-}
-
-
-.title {
-  text-align: center;
-  font-size: 1.8rem;
- 
-  color: var(--black);
-}
-
-
-
-
-.signbutton {
-  margin: 1rem 0;
-}
-
-
-.submitButton:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-
-.errorMessage,
-.successMessage {
-  text-align: center;
-
-  width: 100%;
-  max-width: 400px;
-  font-size: 1rem;
-  color: #ff3333;
-}
-
-.successMessage {
-  color: green;
-}
-
-
-.inputActiontext {
-  font-size: 1rem;
-  color: var(--description);
-  font-weight: 500;
-  text-align: center;
-  margin-top: 10px;
-}
-
-.inputActiontext .signIn {
-  color: #007bff;
-  cursor: pointer;
-}
-
-.inputActiontext .signIn:hover {
-  text-decoration: underline;
-}
-
-
-@media (max-width: 480px) {
-
-  .form {
-    padding: 15px;
-    width: 90%;
-  }
-
-
-  .title {
-    font-size: 1.5rem;
-  }
-
-  .submitButton {
-    font-size: 16px;
-    padding: 10px;
-  }
-}
